Update patch set 1
Patch Set 1:

Kaishu: Thank you for the patch!

Though what we need here is a bit more complicated than this :-)  The keyboard showing or hiding should be handled directly from the JavaScript; and there are some patches near completion that will improve the behavior here (I've CC'd Tamas Zolnai who is working on them).

What I need from you is a bit different: Add a button in the lower toolbar (ideally one that somehow "stands out" - isn't scrolling as the other), and when this one is pressed, the keyboard hides, and _instead_ of the keyboard, the actual sidebar appears.

Ie. want to change where and how the sidebar appears, instead of rolling from the right side, it would appear instead of the keyboard, so that the user can still see the document and operate the sidebar at the same time.

Of course - the real fun will begin when the user clicks into a field that needs a keyboard :-) - for that we'll need to move the sidebar to the document's area; but move it back again instead of the keyboard when the user finishes the typing.

Does that make sense?  Happy to have a call about this if it is still unclear.

Patch-set: 1
Reviewer: Gerrit User 1000068 <1000068@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Style=0