tdf#126971: Fix localisation of tunnelled dialogs in the iOS app
This used to work fine, but at some stage it had regressed. I did not
bother bisecting to find the culprit. Anyway, now the JS part needs to
send a lang parameter in the load message, too, for the corrent
language to be used in the tunnelled dialogs. (Likely it affects also
other things in core.)

Change-Id: I339cddb28357978414669964106cb6fa472c967f