suppress mouseout even when dragging a graphic selection
This fixes the problem when a mouseout event interrupts dragging
of a graphic selection. Not sure if it is the most correct thing
to do, but we can't set "pointer-events" to none on the graphic
selection as this then wouldn't recognise that we want to move
it around.

Change-Id: I42a4785fec0b674d47cd4d7c2977877607b03db8
Reviewed-on: https://gerrit.libreoffice.org/78200
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Tomaž Vajngerl <quikee@gmail.com>
(cherry picked from commit 1e044a40aa9e0b8e2313d4677755c630cb444448)