Make changes to "Resolved Comments" view menu option
The check mark for resolved comments in the view menu
is now dependent on the state of the javascript variable
rather than the other way around. This brings it in line
with how other similar menu options work.

Also fixed ruler so that it looks correct when there are
comments on the document which are all hidden.

Change-Id: I636260da3d053799b53363e9412a0b367e3a9274