tdf#125575: Must reset the "kit" callback to null when closing document
As we register the global "kit" callback with the Document object
pointer as callback data, we must unregister it when the document is
being "unloaded" and the Document object will destruct shortly.
Otherwise the callback might be called for the next document with a
stale pointer.

Relevant only for the iOS app, and presumably the Android one. In
web-based Online, the kit process will exit anyway after the document
is closed. But in an app there is just one process that keeps running
for a long time handling all documents that are edited.

For some reason the crash only showed up when a document was created
from a template, closed, and another document was created from a
template.

Change-Id: Ieb47838a597c5bcab89c1a33d9bc2f2adb57b3b3