tdf#125575: Don't be so eager to clean up too early in the MOBILEAPP case
It's not needed in an app, and causes misbehaviour when the lifecycle
of DocumentBroker objects gets screwed up.

Many of the desperate attempts to "clean up" here and there in the
code are likely relevant only in web-based Online to catch weird error
situations. In an app our objects hopefully have a much simpler
lifecycle.

Change-Id: I788171cd1ca114ecaca276a7f24185ff67b984c7