tdf#127669: Add more export formats for mobile apps
The same formats are offered as for "Download as" in web-based Online.
But in the apps, the menu entry is called "Export as", not "Download
as", because the document is not being edited on some server from
which it would be downloaded.

Change-Id: I5b1bfc3300bae2c213812c2534ae6b583423662c
Reviewed-on: https://gerrit.libreoffice.org/79964
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>
(cherry picked from commit c7bfc05dcb1c475c1bf9a70d39732790793380d9)