loleaflet: improve context toolbars in the iOS App
On double tap on a word, the word is selected and a 'Cut|Copy|Paste'
toolbar is shown.
On double tap on an empty line or after a space a 'Paste' toolbar is
shown and the text cursor is moved to tap point.
On press on a not selected word the text cursor is moved to the touch
point and a 'Paste' toolbar is shown.
On press on a the selected cell a 'Paste' toolbar is shown.
On press on the text cursor handle a 'Paste' toolbar is shown.
The 'Paste' entry is shown only if there is a text content that can be
pasted. An empty toolbar is never shown.
On single tap the context toolbar is removed.
On panning, pinching, dragging the text cursor handle or dragging the
text selection handles a context toolbar (if active) is hidden and
then shown again at the correct position when the action is over.
On typing and on lost focus the context toolbar is removed.
The context toolbar is placed so that to be centered wrt the text
cursor or the north center of the text selection bounds.

L.Control.ContextToolbar has been generalized so that now can be
easily populated with any entry and extended to support any command.

Change-Id: Ia112a491d952ed5b32c6dfbb600fc414ac1181e3