Update patch set 2
Patch Set 2:

it's just that on mobile there a couple of different user patterns for panes and if we want to have a visible top bar then our wizard should be able to transform into a full vertical view when a) dragging the title bar up or/and b) scrolling. Basically is due to the fact that we should be able to see only on "app/title bar" at the time with the exception of a view that can grow to be full vertical. In this case as seen for example in many of the android apps the header changes from white to full color when in fullscreen. Other than that I see no problem in having no color difference since a colored bar in a small pane can become quite heavy and distracting for what it really matters: its content and the document's content

Patch-set: 2
Reviewer: Gerrit User 1002126 <1002126@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0