loeaflet: remove empty HTML elements
The document type is detected when 'doclayerinit' event is fired.
At that point we know the document type so we can remove safely
the empty elements that they are useless to the document and
of course a clean document to debug. I hope in the near future
the server side can remove these unwanted elements,
anyway how is useful an empty html element?

Change-Id: I25f5b8cbcd176c4cc66abcb37b7d0d4b588d1d96