loleaflet: debug: use unique name for document container
I had a not good debugging session trying to find out when
the CSS properties of the document container are changing dynamically.
The best scenario is to use the unique name documentContainer, so we can
track all CSS properties dynamic changes, and it can set breakpoint in a
suspicious change for deeper analysis.

Change-Id: I3accef1c374689f04da336cd0a84047b498540b1