Update patch set 1
Patch Set 1:

> We're trying to reduce the Poco dependency in online. I'd prefer
 > not to add new pieces there if we can at all avoid it =) a single
 > extra print statement is -far- cleaner weight-wise surely.

Well - I don't see it as a new dependency; the XMLWriter is used in the code anyway (being part of the DOMWriter), and we just call its method... which is cleaner in the sense that it would correctly set the used encoding, while writing it manually just hard-codes it. But - no practical problem with any way (since we probably want utf-8 anyway).

Patch-set: 1
Reviewer: Gerrit User 1000994 <1000994@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Code-Style=0