Update patch set 1
Abandoned

Let me abandon this - based on the fact that it turned out to be unnecessary. It was my initial guess that this could be the reason of the XML rejected by SP; but that's not the case: after I figured that it's missing proof-key element that caused the problem, and re-tested the XML to be accepted by SP without the declaration string, let us keep it simple.

Patch-set: 1
Status: abandoned