loleaflet: remove unused loleaflet/debug folder
Long time ago, the leaflet project was imported to be
reused for document tile rending instead of map tiles.
During that time they were added so many commit and the
leaflet source code suffer modifications and mutations like:
new functions are removed dead code (related to map Geo locations).
Now if you try to open some of the files, it will not work,
and to not confuse new contributors, it is preferable to remove all this
unused files.

But if you still interested in how to use all these files,
please go to the original repository
https://github.com/Leaflet/Leaflet.git
and checkout the branch version 0.8.0-dev, you will be happy how all
this file works very well.

However, if you want to debug the new changes, I strongly recommend you
to add

./configure --enable-debug  /* other parameters */

to debug Collabora Online.

Any objections?

Change-Id: I38c85f5716f608390e79edbffecacb3561fda97a