loleaflet: fire 'orientationchange' event
In order to detect if the orientation of the device has changed, it is
required to capture the window resize event. However, there is another
event attached from iframe element called resize-detector. I am not sure
how is used, but I do not want to cause any regression from desktop
version just let's add another event listener and it will be fired on
mobile case.

Change-Id: I54b5ba9239a71216faae4757a0cc7d9826c443ff