service: improve stop and reload
SIGINT is the clean way of stopping
the service, while SIGTERM is be
the faster and rude way. So first
we want to use SIGINT and only when
we timeout do we want SIGTERM to
be used.

Here we ask systemd to stop WSD
by sending SIGINT and wait for 120
seconds. After that timeout, it will
assume the service is hung and needs
to be aborted. We expect that by then
we would have saved all documents and
cleanly exited. 120 seconds should be
sufficient time to save, close, and
upload the open documents.

Change-Id: I08321814db942d89a44cfce42885840f4afa279c