Update patch set 1
Patch Set 1: Code-Review-1

(12 comments)

This is a great start =) lots of promising things here, and looking forward to having a nice REST monitoring end-point. A quick skim / review - I had a couple of questions:
* how is this authenticated ?
* what is the /proc/ regexp search used for ;-)
that I need to investigate more for the next iteration, but - a great start.
Thank Gabriel !

Patch-set: 1
Label: Code-Review=-1