wsd: Destroy ChildProcess instances before exit
ChildProcess instances log in the dtor, and this can
happen after the log subsystem has been destroyed,
because the order of destruction is undefined.
NewChildren is destroyed in exit, at which point
we have no control over the log subsystem.

To avoid this inversion of order, we clear
the NewChildren container before exiting.

Change-Id: I7b1bac3123fd7917616c11ab9674534c9222fff1