Update patch set 1
Patch Set 1:

wrt. the kill - fair enough =)
wrt. the groping around in /proc - it is something we used to do a lot and I really hate it: can we move that code into the AdminModel if you really want it and not put it somewhere where someone else might use it ;-)
wrt. internal binary formats - lets re-factor that in a second step if we ever want it. In general I'm not a huge fan of the pages full of very similar text in AdminMetrics::toString etc. - it all looks very verbose and unwieldy to me =) Perhaps if they were an array of doubles with enums, and an array of names it would look prettier (?) ... currently the sheer density of hard-coded text there is amazing ;-)

Patch-set: 1