Update patch set 1
Patch Set 1:

Unfortunately this does not work :-(  The good news is that this is covered by the fast 'whitebox' unit tests that are directly part of the build - so just do "make -j" and see it fail :-)

I'd like to have the regex here to be std::regex(R"(\n\r)") so that it catches the entire '\n\r' phrase as a delimiter.  Then it still fails - because one of the test cases tests against extra whitespace at the end.

You'll need to extend the for (const auto& token : tokens) to find also the end of the text (ie. the first non-' ' at the end) and extend the token.substr(i) accordingly.

Can you do that please?  The test will help you greatly I hope.

Patch-set: 1
Label: Verified=0