Update patch set 2
Patch Set 2:

> Debian 8 (...) are supported. Debian 8 at least contains
> so old npm and nodejs, that is unusable to build loleaflet.
> Therefore we use nodejs packages from
> https://github.com/nodesource/distributions
> That nodejs package contains npm

In the "Debian ecosystem", the way of handling that stuff required by packages is installed outside the packaging system is to use "equivs" to declare to the packaging system that the "required stuff" is installed.

E.g. if building the loolwsd Debian-format source package requires at least npm version 5, then the correct build-dependency is "npm >= 5". Now, you tell me that this dependency is not satisfiable on Debian 8. So, to build on Debian 8, one would have to either use a Debian-format package backport, or install it outside the packaging system (as you tell me you do) and use equivs to teach the packaging system that npm is installed.

It would be considered a bug of the source package that it doesn't list (all) what it needs (beyond build-essential, which gives you a C compiler, libc development headers, GNU Make, etc).

Patch-set: 2
Reviewer: Gerrit User 1000084 <1000084@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0