Update patch set 2
Patch Set 2:

Mert - thanks for that!

Can you add a short comment above the 'if (...)' explaining why that; like "Don't apply watermark outside of the document area, only the document area has full opacity - the rest (the shadow and the outside area) is partly or fully transparent" - or something to that?

Patch-set: 2