tdf#128468: Make dragging the margins in the ruler work better in iOS app
Note that in this branch it seems even harder to touch a margin handle
in the ruler exactly in the right spot to be able to drag it. You will
have to use your smallest finger and point very exactly. And still it
works only sometimes. (But I did get it to work once, honestly, so
ship it!) This obviously needs to be fixed, but cherry-picking this
commit as is for now.

Several changes:

On iOS, use touch events instead of mouse ones. One difference between
touch and mouse events is that the touchend event naturally does not
carry any position information (as the finger has been lifted from the
screen). So we need to keep track of where the last touchmove event
happened separately in _lastPosition.

Letting go of a margin handle after dragging it did not actually move
the coresponding margin in the document. The fix for this was to catch
the touchend event for the this._rFace element instead of for the
this._map element.

Once a margin had been moved, it could not be moved another time. The
reason for this was that because the document initially is readonly,
the this.options.interactive was set to false (see the
L.control.ruler() call in loleaflet/src/map/Map.js) and never changed
after that. This caused _updateBreakPoints() to pass {perm:'readonly'}
in the call to this._changeInteractions(). The fix was to set
this.options.interactive to true when turning on the interaction.

In theory this same problem probably would happen also in web-based
Online on "mobile" (i.e. phones), but we don't show the ruler there at
all, so the sitaution where a document would be initially read-only
and the ruler would show up doesn't happen.

Reviewed-on: https://gerrit.libreoffice.org/83807
Reviewed-by: Tor Lillqvist <tml@collabora.com>
Tested-by: Tor Lillqvist <tml@collabora.com>
(cherry picked from commit caf37fb293a666b0a82ac8bd31df35b18b45fc01)

Change-Id: I0393704d02a2989b4ea3358dc4bee17e48c16ae6