Improve ruler behaviour when panning by dragging on a touch device
Fire the 'scrolloffset' event also in
L.Control.Scroll._onUpdateScrollOffset(), which gets called when
panning a text document on a touch device.

There are still glitches in updating the ruler when one uses a pinch
gesture to zoom the document. But usually after a short moment the
ruler snaps into correct state.

Change-Id: Ic8dd5139ff49f6fe7961096552f1c7e6cf35356c