tdf#128468: Draw the tab stops
Draw only the user-edited/able ones. Draw all four kinds (left, right,
center, and decimal) the same for now, as a small triangle pointing
up. (Drawn using a CSS trick to get a triangle.) The tab stop
positions aren't actually editable yet, nor is there a way to add a
new tab stop or delete one. Later.

The code uses a CSS custom property, yay, to avoid duplication of a
magic value for the half-width of the triangle.

Change-Id: I862b09091bad03ad63425cc6e6759f3f13174579