leaflet: sidebar scrolling improvements
The browser has to move the sidebar to accomodate
scrolling, and this movement makes the sidebar
overlap with the document (specifically, the
scrollbar of the document) and causes other visual
glitches.

We now display a scrollbar at all times to make
the sidebar width fixed and also to avoid flickering.
This is the only portable/reliable way of avoiding
flickering/jumping-around due to scrollbar showing
and hiding. (Previously we had extra space in the
parent, but at least in Chrome it now always pushes
the canvas left when showing the scrollbar anyway,
so we gain nothing but having a thin margin on the
right, and we still got shifted left/right when
the scrollbar show/hide.)

Change-Id: Icbec841ffa4cec255e59f547530728f076d2877e