android: Ultimately we want to have this working for both.
Currently it can open files both from the ownCloud and Nextcloud
storage, but so far it seems that it cannot save back to ownCloud.

Unfortunately both the libraries have the same namespace, so to have
both the ownCloud's and Nextcloud's library in the same .apk, we'll need
to use some additional tricks.

Change-Id: I7fd647a42d3bd90b6268017ae392d61f67f548c0