leaflet: sidebar is now always active for mobile but visible on demand
Because we can't control the sidebar when in Chart edit
mode, we must have it active before we start Chart editing.
Besides that, on mobile we show/hide the sidebar quite
frequently, and there is no point in activating/deactivating
the sidebar on every tap.

So we have the sidebar enabled in the background at all times.
We then track whether it's visible or not (window.mobileWizard)
and show/hide it. Furthermore, we now handle the sidebar
close notification from Core, which we don't expect, really,
now that we don't close it. However, we need to have a way
to re-activate the sidebar when it is closed. Something
we don't do right now, and we need to tack the Core
sidebar state as we can only toggle the sidebar.

Change-Id: Id7774b84cd0604055c708eb0912e17988f5823d2