calc: tunneled formula bar: skip painting when width is wrong.
In the hope of minimizing flikering effects we skip to paint the
formula bar when the size is not the expected one, on the contrary we
send a resize request with the correct width to the lok core.

We also skip to adjust/resize the formula bar when a sidebar resize
not really change the sidebar width.

Change-Id: I11d343b2ba722f96d3cdba6a2dfb66d85a071c5f