leaflet: fix long-tap in text selection
We only track selections with the corner points.
This makes checks for where the mouse event is
inaccurate, because text selections are rarely
squares. Typically, they are rectangles spanning
the whole width of the line, and so we need to
capture the actual area (which we don't).

Here we assume the text is the full width of the
document, which is true most of the time, so our
check for where we tap.

Change-Id: I79d005aa707e29ba90b82cfcbfef1a3cdb08b629