Update patch set 3
Patch Set 3:

it is empty, since weblate triggers commit immediately for new files - for regular stuff it only commits when there is stuff and squashes the changes. And the timeframe when to commit changes is set to 24 hours, so it is no bug that it doesn't show up yet.

Patch-set: 3
Label: Code-Style=0