Update patch set 1
Patch Set 1:

> Any chance you can review & test this Tamas ?

Review what? It's a quick hack to have something. It's known to be not working in general.
If you would like to include it in the next build, then you can do it. The actual fix would be to get the whole selection area from core, right? So there is nothing to review here.

Patch-set: 1