mobile-wizard: Fix execution of insert header / footer menu's All item
In desktop case, it depends on the checked state whether we insert
or delete a header / footer. However, when deletion is triggered
there is a dialog coming up, what we would like to avoid on mobile.
So on mobile we allow only insertion by now. So we can add
"On:bool=true" param to All menu item.

Change-Id: I8ee764c488563b04226cc53e744034b22c0437ff