Update patch set 2
Patch Set 2:

cc'ing someone only means something when you grant some time to react, esp. on holiday season…

Anyway - was enough for me to improve the process a little/to also enable automatic push notifications back to weblate, so at least updating of the pot files should now be reflected ~ immediately in weblate (it has msgmerge hook enabled to automatically updates po files to match the pot) - so far it only bothered to check when it had something to push. While it wouldn't have helped in this case, it helps with weblate accumulating changes in its review request that already had been pushed, resulting in push to gerrit to fail because the review is closed.

So not all bad with this change despite the additional work :-)

Patch-set: 2