Update patch set 1
Patch Set 1:

Can you confirm that you checked this makes sense for the web UI and menus there, and that this is consistent. In my mobile / writer - I see Insert->"Insert Shape" too.

In general we want a complete solution that is tested in both web and mobile UI's to look good; but perhaps I'm mis-understanding what unocommands.js is doing =)

HTH.

Patch-set: 1
Label: Code-Style=0