leaflet: deduplicate dialog paint requests
The sidebar and input-bar can get a very high rate of invalidations
when the user is typing. In many cases this is highly redundant and
noisy. We minimize this noise by deduplicating identical paint
requests by delaying the requests by 250ms. If a new region needs
painting, we flush the previous one and queue up the new message.

The delay is small enough to be acceptable, but large enough
to reduce redundant render requests significantly.

Change-Id: If4c71c2cc6f5c0b3daec15cd4da4c6fbe437b11c