Update patch set 1
Patch Set 1:

Ash: Are you sure about the 250ms?  That is _ages_ in the human's perception of interactivity; so when the user will type in the calc input bar or in the sidebar's input fields, it will feel terribly sluggish.

Also it feels curious to do the deduplication in JS, when we do all the other deduplication in the message queue?

So - if deduplicating, please let's move it to the message queue.  And then as the next step, it would be great to transit the dialog/sidabar/calc-input-line to the same system we are using for tiles: Keep track of the areas in kit/wsd (not 100% sure without further looking where it is), and implement pushing that & sending back the ACK in JS; and make sure that just 2 paints are in-flight, and before we get the the ACK, just grow the area we should paint.

Is that possible?

Patch-set: 1