Update patch set 1
Patch Set 1:

Let me clarify. The 250ms is perfectly fine for updating the sidebar/dialogs, as there is already noticable delay when updating the toolbar/menus when changing context (it's much higher latency), and this doesn't make it much worse. We can reduce it of course, if we need faster sidebar updates (but practically 250ms is on the edge of what an average person can notice, and the sidebar isn't as critical as typing, but it *can* slow typing down if we invalidate/paint it too frequently, which is what I'm reducing here).

Regarding where to do the dedup, I agree that ideally we shouldn't emit these invalidations in the first place, but once we do, there is nothing preventing the client from requesting unnecessary paint requests, which is what I'm reducing here.

With this patch, I can skip 100s of unnecessary sidebar updates when typing, which really helps performance and latency. But I agree we should ideally track what area each client is notified to be invalid and refrain from invalidating that area until/unless the client renders it (then we send invalidations _after_ rendering, if necessary).

Patch-set: 1