Update patch set 1
Patch Set 1:

To see the benefit of this patch, have two views with sidebar and type in both (enable auto-typing in one, and type manually in the other to see how sluggish it can get due to the sidebar). Disable the sidebar in both and try again to see the overhead of the sidebar invalidation. With this patch, the performance of simultaneous typing is noticeably better. (With a single user typing the sidebar is not such a big issue, because we have bandwidth.)

Patch-set: 1
Reviewer: Gerrit User 1000871 <1000871@6622c302-437a-49ad-a91d-55ba7dcf1615>
Label: Verified=0