leaflet: don't expose the textInput object via getter
Contain the textInput instance in map and use
the available focus() and blur() functions instead.
Ideally these functions would be renamed to something
more accurate, but that's a big and risky change.

Note that there are many cases where the textInput is
still accessed directly. These are harder to encapsulate
just yet, but we should move in that direction.

Change-Id: Ibc14fcbde14b54a9633cf534e11072f6fd383403