android: improve clipboard, inter-document copy/paste possible
In andorid app, we just supported html or text to be copied between
applications (inside the document the copy/paste is handeled by
the core). The biggest problem with this is that copying between
documents doesn't work or works just as good as our HTML support
is. With this change, we get all the clipboard content after copy
and serialize it to a file into the cache and if we then paste in
another document, we read the content of the file and paste
without any loss of fidelity.

Change-Id: I3501045ff2e815a7e2f8c6ecce5255511c256ca6