Show the document file name in the mobile apps, too
Show it in the same text input field that is used in the WOPI-like
case. Make the text input field non-editable and not even focusable.

Show the file name as it should be displayed to the user. I.e. reverse
two layers of URL encoding. At least for the iOS app this is
necessary.

Change-Id: Ifc3c8c0907178ecb419f3108ffc558a05a16313d