android: clipboard is sync. so exec. in the thread is not needed
As we changed the clipboard to be synchronous now, it's not needed
to execute copy/cut/paste actions in a separate thread anymore.

Also change WebView message interception, so we can execute
actions after a we send WebView a message. This is needed when
we copy/cut something and need to push the content to the android
clipboard and clipboard file.

Change-Id: I08b6ee55ca8bd7b958d348e257c19395ea39ac80