Update patch set 2
Patch Set 2:

There was no need to patch core.git as the comment position send by it are correct. The issue is that online did not request for it when zoom/view-reset happens. I have tested ok with comments at the bottom of the sheet, and with two users with different zoom levels.

Patch-set: 2